[Move `@kbn/config-schema` to server] `discover`
