+++ src/plugins/discover/server/config.ts
@@ -7,7 +7,20 @@
  */
 
 import { PluginConfigDescriptor } from '@kbn/core-plugins-server';
-import { configSchema, type ConfigSchema } from '../common/config';
+import { schema, TypeOf } from '@kbn/config-schema';
+
+export const configSchema = schema.object({
+  enableUiSettingsValidations: schema.boolean({ defaultValue: false }),
+  experimental: schema.maybe(
+    schema.object({
+      ruleFormV2Enabled: schema.maybe(schema.boolean({ defaultValue: false })),
+      enabledProfiles: schema.maybe(schema.arrayOf(schema.string(), { defaultValue: [] })),
+    })
+  ),
+});
+
+export type ConfigSchema = TypeOf<typeof configSchema>;
+export type ExperimentalFeatures = NonNullable<ConfigSchema['experimental']>;
 
 export const config: PluginConfigDescriptor<ConfigSchema> = {
   schema: configSchema,
